+++ indico/modules/events/roles/__init__.py
@@ -24,3 +24,9 @@ def _sidemenu_items(sender, event, **kwargs):
     if event.can_manage(session.user):
         roles_section = 'organization' if event.type == 'conference' else 'advanced'
         return SideMenuItem('roles', _('Roles Setup'), url_for('event_roles.manage', event), section=roles_section)
+
+
+@signals.event_management.get_cloners.connect
+def _get_event_roles_cloner(sender, **kwargs):
+    from indico.modules.events.roles.clone import EventRoleCloner
+    return EventRoleCloner